+++ game.py
@@ -10,6 +10,8 @@ import numpy as np
 import threading
 import time
 from map import Map
+from snake import Snake
+from config import Config
 
 
 class Settings:
@@ -21,138 +23,6 @@ class Settings:
         # self.width = 52
         # self.height = 52
         # self.rect_len = 15
-
-class Snake:
-    def __init__(self):
-        
-        self.image_up = pygame.image.load('images/head_up.bmp')
-
-        self.tail_up = pygame.image.load('images/tail_up.bmp')
-        
-        self.image_body = pygame.image.load('images/body.bmp')
-        self.image_body_s = pygame.image.load('images/body_s.bmp')
-        self.image_body_c = pygame.image.load('images/body_c.bmp')
-        self.image_space = pygame.image.load('images/space.bmp')
-        
-        #self.facing = "right"
-        self.initialize()
-
-    def initialize(self, mapdir=None):
-        if mapdir is not None:
-            with open("levels/" + mapdir + "/snake.txt", "r") as f:
-                t = f.readlines()
-            self.segments = []
-            self.segmentd = []
-            for pair in t:
-                mid = pair[:-1].split(",")
-                self.segments.append([int(mid[0]), int(mid[1])])
-                self.segmentd.append([int(mid[2]), int(mid[3])])
-            dire = self.segments[1]
-            if dire == [-1, 0]:
-                self.facing = "right"
-            if dire == [0, 1]:
-                self.facing = "up"
-            if dire == [1, 0]:
-                self.facing = "left"
-            if dire == [0, -1]:
-                self.facing = "down"
-        else: #default to fallback to
-            self.segments = [[6, 6], [-1, 0], [-1, 0], [-1, 0]]
-            self.segmentd = [[6, 6], [5, 6], [4, 6], [3, 6]]
-            self.facing = "right"
-        #this is so wacky
-        self.score = 0
-
-    def blit_body(self, loc, cur, next, screen, size):
-        x, y = loc[0]*size, loc[1]*size
-        direction = [2*cur[0] + next[0], 2*cur[1] + next[1]]
-        if direction[0] == 0:
-            screen.blit(self.image_body_s, (x, y))
-        elif direction[1] == 0:
-            screen.blit(pygame.transform.rotate(self.image_body_s, 90), (x, y))
-        elif direction == [2, 1] or direction == [-1, -2]:
-            screen.blit(self.image_body_c, (x, y))
-        elif direction == [-2, 1] or direction == [1, -2]:
-            screen.blit(pygame.transform.rotate(self.image_body_c, 90), (x, y))
-        elif direction == [-2, -1] or direction == [1, 2]:
-            screen.blit(pygame.transform.rotate(self.image_body_c, 180), (x, y))
-        elif direction == [2, -1] or direction == [-1, 2]:
-            screen.blit(pygame.transform.rotate(self.image_body_c, 270), (x, y))
-        else:
-            screen.blit(self.image_right, (x, y))  
-
-    def blit_head(self, loc, dire, screen, size):
-        x, y = loc[0]*size, loc[1]*size
-
-        if dire == [0, 1]:
-            screen.blit(self.image_up, (x, y))
-        elif dire == [0, -1]:
-            screen.blit(pygame.transform.rotate(self.image_up, 180), (x, y))  
-        elif dire == [1, 0]:
-            screen.blit(pygame.transform.rotate(self.image_up, 90), (x, y)) 
-        else:
-            screen.blit(pygame.transform.rotate(self.image_up, 270), (x, y))
-
-
-    def blit_tail(self, x, y, screen, size):
-        #tail_direction = [self.segments[-2][i] - self.segments[-1][i] for i in range(2)]
-        tail_direction = self.segments[-1]
-        x *= size
-        y *= size
-        print(tail_direction)
-        print(self.segments)
-        print(self.segmentd)
-        #ls[big iterator][small iterator]
-        
-        if tail_direction == [0, 1]:
-            screen.blit(self.tail_up, (x, y))
-        elif tail_direction == [0, -1]:
-            screen.blit(pygame.transform.rotate(self.tail_up, 180), (x, y))    
-        elif tail_direction == [1, 0]:
-            screen.blit(pygame.transform.rotate(self.tail_up, 90), (x, y))
-        else:
-            screen.blit(pygame.transform.rotate(self.tail_up, 270), (x, y)) 
-
-    def blit(self, rect_len, screen, last=None):
-        for index, position in enumerate(self.segments[1:-1]):
-            self.blit_body(self.segmentd[index + 1], position, self.segments[index + 2], screen, rect_len)
-        self.blit_tail(self.segmentd[-1][0], self.segmentd[-1][1], screen, rect_len)
-        if isinstance(last, list):
-            #this only outputs regular spaces, since the features in it get added later than the snake!
-            screen.blit(self.image_space, (last[0]*rect_len, last[1]*rect_len))
-        self.blit_head(self.segmentd[0], self.segments[1], screen, rect_len)
-        
-    def update(self, straw):
-        pos = [0, 0]
-        if self.facing == 'right':
-            pos[0] += 1
-        if self.facing == 'left':
-            pos[0] -= 1
-        if self.facing == 'up':
-            pos[1] -= 1
-        if self.facing == 'down':
-            pos[1] += 1
-        headpos = [self.segments[0][0] + pos[0], self.segments[0][1] + pos[1]]
-        #convert headpos for wrap or solid tile collision
-
-        #check for body collision, if yes the snake doesnt move forward.
-        if headpos in self.segmentd[:-1]:
-            return -1, []
-        self.segments.insert(0, headpos)
-        self.segmentd.insert(0, headpos)
-
-        #check for strawberry at head pos now.
-        if self.segments[0] == straw.position:
-            straw.random_pos(self)
-            reward = 1
-            self.score += 1
-            last_tail = [-1, -1]
-        else:
-            self.segments.pop()
-            #unblit tail
-            last_tail = self.segmentd.pop()
-        self.segments[1] = [-pos[0], -pos[1]]
-        return 0, last_tail
 
 class Strawberry():
     def __init__(self, settings):
@@ -193,14 +63,12 @@ class Game:
                           3 : 'right'}   
         self.tile_img = pygame.image.load('images/tile.bmp')
         self.space_img = pygame.image.load('images/space.bmp')
-    '''
-    def load_map(self, mapfile)
-        self.map = Map(mapfile)
-    '''
 
-    def restart_game(self, mapdir):
+    def restart_game(self, mapdir): 
+        #set config. This has a bunch of options that control stuff.
+        self.config = Config(mapdir)
         #set map
-        self.map = Map(mapdir)
+        self.map = Map(mapdir=mapdir, config=self.config.settings)
         #set snake
         self.snake.initialize(mapdir)
         #set stawberry if it exists.
@@ -240,22 +108,7 @@ class Game:
             self.snake.facing = change_direction
 
         state, replace = self.snake.update(pos)
-        #moves maybe? Yes.
-        '''
-        if self.snake.position == self.strawberry.position:
-            self.strawberry.random_pos(self.snake)
-            reward = 1
-            self.snake.score += 1
-        else:
-            #self.snake.segments.pop()
-            reward = 0
-        #point and eating        
-        
-        if self.game_end():
-            return -1
-        return 0            
-        #return reward
-        '''
+
         if state == -1:
             return -1
         if replace and replace != [-1, -1]:
@@ -277,16 +130,6 @@ class Game:
         font = pygame.font.SysFont(None, 25)
         text = font.render('Score: ' + str(self.snake.score), True, color)
         screen.blit(text, (0, 0))
-
-    def blit_border(self, screen):
-        # using one of the food tiles as a wall for the time being 
-        # no changes yet to make it so that the player dies when it hits the walls- 
-        # this would be better implemented with the use of a map class
-        for i in range(1, 28): 
-            screen.blit(pygame.image.load('images/food8.bmp'), (0, 15*i))
-            screen.blit(pygame.image.load('images/food8.bmp'), (15*i, 15))
-            screen.blit(pygame.image.load('images/food8.bmp'), (405, 15*i))
-            screen.blit(pygame.image.load('images/food8.bmp'), (15*i, 405))
 
     def blit_map(self, screen): 
         for i in range(0, 28):
