+++ levels/1-1/config.txt
@@ -1,0 +1,3 @@
+what:=2
+how:=3
+why:=4
